* contrib/cc-with-tweaks.sh (-p): Handle no dwo files.

+++ gdb/contrib/cc-with-tweaks.sh
@@ -179,10 +179,13 @@ fi
 if [ "$want_dwp" = true ]; then
     dwo_files=$($READELF -wi "${output_file}" | grep _dwo_name | \
 	sed -e 's/^.*: //' | sort | uniq)
-    $DWP -o "${output_file}.dwp" ${dwo_files} > /dev/null
-    rc=$?
-    [ $rc != 0 ] && exit $rc
-    rm -f ${dwo_files}
+    rc=0
+    if [ -n "$dwo_files" ]; then
+	$DWP -o "${output_file}.dwp" ${dwo_files} > /dev/null
+	rc=$?
+	[ $rc != 0 ] && exit $rc
+	rm -f ${dwo_files}
+    fi
 fi
 
 rm -f "$index_file"
